pyproject.toml
[tool.poetry]
name = "fastapi-contacts"
version = "0.1.0"
description = "FastAPI Contacts Management Application"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.104.1"
uvicorn = "0.24.0"
sqlalchemy = "2.0.23"
psycopg2-binary = "2.9.9"
pydantic = "^2.10.1"
python-dotenv = "1.0.0"
alembic = "1.12.1"
email-validator = "^2.2.0"
cloudinary = "^1.44.0"
fastapi-mail = "^1.4.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
python-multipart = "^0.0.9"
slowapi = "^0.1.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
